refactor: 장바구니 아이템을 추가하는 로직 개선
